リファクタリングofBallbot
