+++ src/components/BallBot/constants.ts
@@ -1,0 +1,128 @@
+// Animation timing constants
+export const ANIMATION_CONSTANTS = {
+  // State transition speeds
+  THINKING_SPEED_SLOW: 0.02,
+  THINKING_SPEED_NORMAL: 0.03,
+  THINKING_SPEED_FAST: 0.06,
+  SPEAKING_SPEED_SLOW: 0.08,
+  SPEAKING_SPEED_NORMAL: 0.10,
+  SPEAKING_SPEED_FAST: 0.15,
+  SPEAKING_FADEOUT_SLOW: 0.02,
+  SPEAKING_FADEOUT_NORMAL: 0.04,
+  TRANSITION_SPEED: 0.06,
+  
+  // Transition thresholds
+  THINKING_TRANSITION_THRESHOLD: 0.3,
+  SPEAKING_TRANSITION_THRESHOLD: 0.5,
+  SPEAKING_AFTEREFFECT_THRESHOLD: 0.5,
+  
+  // Animation durations
+  TRANSITION_DURATION_FRAMES: 210, // ~3.5 seconds at 60fps
+  MIN_SPEAKING_DURATION: 7, // seconds
+  CHAR_TO_DURATION_RATIO: 0.6, // seconds per character
+  FPS: 60,
+  
+  // Animation phases
+  PHASE_1_END: 0.3,
+  PHASE_2_END: 0.5,
+  PHASE_3_END: 0.8,
+  SPEAKING_FADE_START: 0.9,
+};
+
+// Visual effect constants
+export const VISUAL_CONSTANTS = {
+  // Wave effects
+  WAVE_TIME_SCALE: 0.3,
+  WAVE_X_FREQUENCY: 0.4,
+  WAVE_Y_FREQUENCY: 0.3,
+  WAVE_Z_FREQUENCY: 0.25,
+  WAVE_X_AMPLITUDE: 0.005,
+  WAVE_Y_AMPLITUDE: 0.007,
+  WAVE_Z_AMPLITUDE: 0.005,
+  
+  // Audio effects
+  AUDIO_AMPLITUDE_SCALE: 0.02,
+  AUDIO_SCALE_FACTOR: 0.1,
+  
+  // Thinking effects
+  THINKING_PULSE_BASE_FREQ: 3.0,
+  THINKING_PULSE_VARIATION: 1.0,
+  THINKING_PULSE_AMPLITUDE: 0.015,
+  THINKING_PULSE_SECONDARY: 0.01,
+  THINKING_SCALE_FACTOR: 0.05,
+  THINKING_VIBRATION: 0.01,
+  
+  // Speaking effects
+  SPEAKING_WAVE_FREQUENCY: 1.5,
+  SPEAKING_RIPPLE_SCALE: 8.0,
+  SPEAKING_RIPPLE_AMPLITUDE: 0.02,
+  SPEAKING_RIPPLE_SECONDARY: 0.015,
+  SPEAKING_SCALE_FACTOR: 0.1,
+  
+  // Energy flow effects
+  ENERGY_FLOW_AMPLITUDE: 0.05,
+  ENERGY_FLOW_FREQUENCY: 3.0,
+  
+  // Float animation
+  FLOAT_Y_SPEED: 0.2,
+  FLOAT_X_SPEED: 0.15,
+  FLOAT_Z_SPEED: 0.175,
+  FLOAT_Y_AMPLITUDE: 0.05,
+  FLOAT_X_AMPLITUDE: 0.025,
+  FLOAT_Z_AMPLITUDE: 0.04,
+  
+  // Rotation
+  ROTATION_BASE_SPEED: 0.02,
+  ROTATION_THINKING_SPEED: 1.5,
+  ROTATION_X_AMPLITUDE: 0.01,
+  ROTATION_THINKING_AMPLITUDE: 0.05,
+  ROTATION_SPEAKING_SCALE: 0.1,
+};
+
+// Responsive scale factors
+export const RESPONSIVE_SCALES = {
+  MOBILE: 1.6,
+  TABLET: 1.5,
+  DESKTOP: 1.3,
+};
+
+// Breakpoints
+export const BREAKPOINTS = {
+  MOBILE_MAX: 768,
+  TABLET_MAX: 1024,
+};
+
+// Shader color constants (as HSL values)
+export const COLOR_CONSTANTS = {
+  // Base colors
+  THINKING_HUE: 0.5, // Cyan
+  THINKING_SATURATION: 0.9,
+  THINKING_LIGHTNESS: 0.75,
+  
+  SPEAKING_HUE_BASE: 0.08, // Gold
+  SPEAKING_HUE_VARIATION: 0.1,
+  SPEAKING_SATURATION: 0.85,
+  SPEAKING_LIGHTNESS: 0.7,
+  
+  RAINBOW_SATURATION: 0.8,
+  RAINBOW_LIGHTNESS: 0.7,
+  
+  // Alpha values
+  ALPHA_MIN: 0.1,
+  ALPHA_MAX: 0.7,
+  THINKING_ALPHA_ADD: 0.3,
+  SPEAKING_ALPHA_ADD: 0.4,
+  
+  // Rim lighting
+  RIM_MIN: 0.3,
+  RIM_MAX: 1.2,
+  STATE_RIM_BLEND: 0.7,
+  
+  // Energy pulse
+  THINKING_PULSE_AMPLITUDE: 0.1,
+  THINKING_PULSE_SECONDARY: 0.05,
+  SPEAKING_PULSE_BASE: 0.2,
+  SPEAKING_PULSE_AUDIO_SCALE: 0.4,
+  ENERGY_PULSE_AMPLITUDE: 0.15,
+  ENERGY_PULSE_SECONDARY: 0.1,
+};
